# backend/api/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação de API da aplicação Intellecta",
        "version": "0.1"
    },
    "paths": {
...
        "/institutions/{institution_id}/classes/{class_id}/users/{class_users_id}": {
            "get": {
                "tags": [
                    "Class Users"
                ],
                "summary": "Get a specific user in a class",
                "operationId": "getClassUserById",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_users_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class user details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassUserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found in class"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Class Users"
                ],
                "summary": "Remove a user from a class",
                "operationId": "removeClassUser",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_users_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found in class"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/institutions/{institution_id}/users/{institution_user_id}": {
            "delete": {
                "tags": [
                    "Usuários da Instituição"
                ],
                "summary": "Remover usuário da instituição",
                "description": "Remove um usuário da instituição (apenas administradores ou o próprio usuário)",
                "operationId": "removeUser",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "institution_user_id",
                        "in": "path",
                        "description": "ID do vínculo usuário-instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário removido da instituição com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - apenas administradores ou o próprio usuário podem remover",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vínculo usuário-instituição não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao remover usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/institutions/{institution_id}/events/{event_id}": {
            "get": {
                "tags": [
                    "Eventos Institucionais"
                ],
                "summary": "Obter detalhes de um evento institucional",
                "description": "Retorna os detalhes completos de um evento institucional específico",
                "operationId": "getInstitutionalEvent",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID do evento institucional",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstitutionalEventResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento ou instituição não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Eventos Institucionais"
                ],
                "summary": "Atualizar evento institucional",
                "description": "Atualiza um evento existente (requer privilégios de admin)",
                "operationId": "updateInstitutionalEvent",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID do evento institucional",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do evento a serem atualizados",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInstitutionalEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InstitutionalEventResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento ou instituição não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Eventos Institucionais"
                ],
                "summary": "Excluir evento institucional",
                "description": "Remove um evento institucional (requer privilégios de admin)",
                "operationId": "deleteInstitutionalEvent",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "description": "ID da instituição",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "description": "ID do evento institucional",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Institutional event deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento ou instituição não encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/events": {
            "get": {
                "tags": [
                    "Eventos de Usuário"
                ],
                "summary": "Listar eventos do usuário",
                "description": "Retorna todos os eventos associados ao usuário autenticado",
                "operationId": "getUserEvents",
                "responses": {
                    "200": {
                        "description": "Lista de eventos do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserEventResponse"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum evento encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Eventos de Usuário"
                ],
                "summary": "Criar novo evento",
                "description": "Cria um novo evento associado ao usuário autenticado",
                "operationId": "createUserEvent",
                "requestBody": {
                    "description": "Dados do evento a ser criado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserEventCreatedResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClassUserResponse": {
                "properties": {
                    "class_users_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "joined_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "class_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    },
                    "code": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "InstitutionalEventResponse": {
                "properties": {
                    "institutional_event_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "institution_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "event_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "event_type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UpdateInstitutionalEventRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3,
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "event_type": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserEventResponse": {
                "properties": {
                    "user_event_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "event_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}